requirements.txt
beautifulsoup4==4.10.0
numpy==1.21.5
Pillow==9.0.0
requests==2.26.0
scikit-learn==0.24.2
nltk==3.6.5
streamlit==1.5.0
